tests: add checkers for file stat and read
This just removes some extra boiler plate for error checking things when
reading or stating a file in a build result.
